+++ src/neuroconv/datainterfaces/ecephys/axona/axonadatainterface.py
@@ -33,7 +33,7 @@ class AxonaRecordingInterface(BaseRecordingExtractorInterface):
         tetrode_id = self.recording_extractor.get_property("tetrode_id")
         self.recording_extractor.set_channel_groups(tetrode_id)
 
-    def extract_nwb_file_metadata(self):
+    def extract_nwb_file_metadata(self) -> dict:
         raw_annotations = self.recording_extractor.neo_reader.raw_annotations
         session_start_time = raw_annotations["blocks"][0]["segments"][0]["rec_datetime"]
         session_description = self.metadata_in_set_file["comments"]
@@ -50,7 +50,7 @@ class AxonaRecordingInterface(BaseRecordingExtractorInterface):
 
         return nwbfile_metadata
 
-    def extract_ecephys_metadata(self):
+    def extract_ecephys_metadata(self) -> dict:
         unique_elec_group_names = set(self.recording_extractor.get_channel_groups())
         sw_version = self.metadata_in_set_file["sw_version"]
         description = f"Axona DacqUSB, sw_version={sw_version}"
@@ -92,7 +92,7 @@ class AxonaUnitRecordingInterface(AxonaRecordingInterface):
     """Primary data interface class for converting a AxonaRecordingExtractor"""
 
     @classmethod
-    def get_source_schema(cls):
+    def get_source_schema(cls) -> dict:
         return dict(
             required=["file_path"],
             properties=dict(
@@ -115,7 +115,7 @@ class AxonaLFPDataInterface(BaseLFPExtractorInterface):
     ExtractorName = "NumpyRecording"
 
     @classmethod
-    def get_source_schema(cls):
+    def get_source_schema(cls) -> dict:
         return dict(
             required=["file_path"],
             properties=dict(file_path=dict(type="string")),
@@ -135,7 +135,7 @@ class AxonaPositionDataInterface(BaseDataInterface):
     """Primary data interface class for converting Axona position data"""
 
     @classmethod
-    def get_source_schema(cls):
+    def get_source_schema(cls) -> dict:
         return get_schema_from_method_signature(cls.__init__)
 
     def __init__(self, file_path: str):