+++ src/neuroconv/datainterfaces/ecephys/neuroscope/neuroscope_utils.py
@@ -6,7 +6,7 @@ from dateutil import parser
 from ....tools import get_package
 
 
-def get_xml_file_path(data_file_path: str):
+def get_xml_file_path(data_file_path: str) -> str:
     """
     Infer the xml_file_path from the data_file_path (.dat or .eeg).