+++ src/neuroconv/datainterfaces/icephys/abf/abfdatainterface.py
@@ -26,7 +26,7 @@ class AbfInterface(BaseIcephysInterface):
     ExtractorName = "AxonIO"
 
     @classmethod
-    def get_source_schema(cls):
+    def get_source_schema(cls) -> dict:
         source_schema = super().get_source_schema()
         source_schema["properties"]["file_paths"] = dict(
             type="array",